added matching pw login in signup page

+++ src/components/partials/Signup.js
@@ -21,21 +21,34 @@ const Signup = (props) => {
     const [name, setName] = useState('');
     const [email, setEmail] = useState('');
     const [password, setPassword] = useState('');
+    const [confirmPassword, setConfirmPassword] = useState('');
+    const [matchPw, setMatchPw] = useState(true);
     // TODO: add from password verification
     const [redirect, setRedirect] = useState(false);
 
     const handleSubmit = e => {
         e.preventDefault();
-        axios.post(
-            `${process.env.REACT_APP_SERVER_URL}/api/signup`,
-            { name, email, password }
-        ).then(response => {
-            console.log(response.data)
-            localStorage.setItem('jwtToken', response.data.token); 
-            setAuthToken(response.data.token);
-            props.handleAuth(response.data.user);
-            setRedirect(true);
-        }).catch(err => console.log(`😖 error`, err));
+        // confirm password
+        if (password == confirmPassword) {
+            setMatchPw(true);
+            // sending data
+            e.preventDefault();
+            axios.post(
+                `${process.env.REACT_APP_SERVER_URL}/api/signup`,
+                { name, email, password }
+            ).then(response => {
+                console.log(response.data)
+                localStorage.setItem('jwtToken', response.data.token); 
+                setAuthToken(response.data.token);
+                props.handleAuth(response.data.user);
+                setRedirect(true);
+            }).catch(err => console.log(`😖 error`, err));
+        } else {
+            // change form entry
+            console.log('pw not matching')
+            setMatchPw(false);
+        }
+
     }
 
     // material-ui styles
@@ -75,7 +88,6 @@ const Signup = (props) => {
     }));
 
     const classes = useStyles();
-//TODO logo not loading
     if (redirect) return <Redirect to='/preferences' />
     return (
         <section>
@@ -109,19 +121,67 @@ const Signup = (props) => {
                                             value={email}
                                             onChange={e => setEmail(e.target.value)}
                                         />
-                                        <TextField
-                                            hintText="Password"
-                                            floatingLabelText="Password"
-                                            required
-                                            type="password"
-                                            id="password"
-                                            label="password"
-                                            placeholder="password"
-                                            variant="outlined"
-                                            color="secondary"
-                                            value={password}
-                                            onChange={e => setPassword(e.target.value)}
-                                        />
+                                        {matchPw ?
+                                        <>
+                                            <TextField
+                                                hintText="Password"
+                                                floatingLabelText="Password"
+                                                required
+                                                type="password"
+                                                id="password"
+                                                label="Password"
+                                                placeholder="Password"
+                                                variant="outlined"
+                                                color="secondary"
+                                                value={password}
+                                                onChange={e => setPassword(e.target.value)}
+                                            />
+                                            <TextField
+                                                hintText="Confirm Password"
+                                                floatingLabelText="Confirm Password"
+                                                required
+                                                type="password"
+                                                id="password-confirm"
+                                                label="Confirm Password"
+                                                placeholder="Confirm Password"
+                                                variant="outlined"
+                                                color="secondary"
+                                                value={confirmPassword}
+                                                onChange={e => setConfirmPassword(e.target.value)}
+                                            />
+                                        </>
+                                        :
+                                        <>
+                                            <TextField
+                                                hintText="Password"
+                                                floatingLabelText="Password"
+                                                required
+                                                error
+                                                type="password"
+                                                id="password"
+                                                label="Password"
+                                                placeholder="Password"
+                                                variant="outlined"
+                                                color="secondary"
+                                                value={password}
+                                                onChange={e => setPassword(e.target.value)}
+                                            />
+                                            <TextField
+                                                hintText="Confirm Password"
+                                                floatingLabelText="Confirm Password"
+                                                required
+                                                error
+                                                type="password"
+                                                id="password-confirm"
+                                                label="Confirm Password"
+                                                placeholder="Confirm Password"
+                                                variant="outlined"
+                                                color="secondary"
+                                                value={confirmPassword}
+                                                onChange={e => setConfirmPassword(e.target.value)}
+                                            />
+                                        </>
+                                        }
                                     </div>
                                     <Button className={classes.linkButton} variant="contained" color="secondary" type="submit">SignUp</Button>
                                 </form>
